Migrate WinPE portion to JScript
VBScript is deprecated in Windows and removed in the September 2023 Windows PE addon.  Although VBScript could be patched back in as of November 2023, it's quicker to update this project to use JScript, which still works.

Powershell is in the future, but that involves adding 150 MB to your 300 MB boot image, so let's avoid that for as long as possible.